tfrecord.py moved to tensorflow_backend.py
