--- keras_contrib/datasets/__init__.py
@@ -1,2 +1,0 @@
-from __future__ import absolute_import
-from .tfrecord import *
